swat compiler warnings in layout.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/10175


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@38759 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/layout/IndicRearrangementProcessor.cpp
@@ -1,6 +1,6 @@
 /*
  *
- * (C) Copyright IBM Corp. 1998-2015 - All Rights Reserved
+ * (C) Copyright IBM Corp. 1998-2016 - All Rights Reserved
  *
  */
 
@@ -94,7 +94,7 @@ void IndicRearrangementProcessor::doRearrangementAction(LEGlyphStorage &glyphSto
         ia = glyphStorage.getCharIndex(firstGlyph, success);
         x = firstGlyph + 1;
 
-        while (x <= lastGlyph) {
+        while (x <= (int32_t)lastGlyph) {
             glyphStorage[x - 1] = glyphStorage[x];
             ix = glyphStorage.getCharIndex(x, success);
             glyphStorage.setCharIndex(x - 1, ix, success);
@@ -115,7 +115,7 @@ void IndicRearrangementProcessor::doRearrangementAction(LEGlyphStorage &glyphSto
         id = glyphStorage.getCharIndex(lastGlyph, success);
         x = lastGlyph - 1;
 
-        while (x >= firstGlyph) {
+        while (x >= (int32_t)firstGlyph) {
             glyphStorage[x + 1] = glyphStorage[x];
             ix = glyphStorage.getCharIndex(x, success);
             glyphStorage.setCharIndex(x + 1, ix, success);
@@ -150,7 +150,7 @@ void IndicRearrangementProcessor::doRearrangementAction(LEGlyphStorage &glyphSto
         ib = glyphStorage.getCharIndex(firstGlyph + 1, success);
         x = firstGlyph + 2;
 
-        while (x <= lastGlyph) {
+        while (x <= (int32_t)lastGlyph) {
             glyphStorage[x - 2] = glyphStorage[x];
             ix = glyphStorage.getCharIndex(x, success);
             glyphStorage.setCharIndex(x - 2, ix, success);
@@ -176,7 +176,7 @@ void IndicRearrangementProcessor::doRearrangementAction(LEGlyphStorage &glyphSto
         ib = glyphStorage.getCharIndex(firstGlyph + 1, success);
         x = firstGlyph + 2;
 
-        while (x <= lastGlyph) {
+        while (x <= (int32_t)lastGlyph) {
             glyphStorage[x - 2] = glyphStorage[x];
             ix = glyphStorage.getCharIndex(x, success);
             glyphStorage.setCharIndex(x - 2, ix, success);
@@ -202,7 +202,7 @@ void IndicRearrangementProcessor::doRearrangementAction(LEGlyphStorage &glyphSto
         id = glyphStorage.getCharIndex(lastGlyph, success);
         x = lastGlyph - 2;
 
-        while (x >= firstGlyph) {
+        while (x >= (int32_t)firstGlyph) {
             glyphStorage[x + 2] = glyphStorage[x];
             ix = glyphStorage.getCharIndex(x, success);
             glyphStorage.setCharIndex(x + 2, ix, success);
@@ -228,7 +228,7 @@ void IndicRearrangementProcessor::doRearrangementAction(LEGlyphStorage &glyphSto
         id = glyphStorage.getCharIndex(lastGlyph, success);
         x = lastGlyph - 2;
 
-        while (x >= firstGlyph) {
+        while (x >= (int32_t)firstGlyph) {
             glyphStorage[x + 2] = glyphStorage[x];
             ix = glyphStorage.getCharIndex(x, success);
             glyphStorage.setCharIndex(x + 2, ix, success);
@@ -256,7 +256,7 @@ void IndicRearrangementProcessor::doRearrangementAction(LEGlyphStorage &glyphSto
         id = glyphStorage.getCharIndex(lastGlyph, success);
         x = lastGlyph - 2;
 
-        while (x > firstGlyph) {
+        while (x > (int32_t)firstGlyph) {
             glyphStorage[x + 1] = glyphStorage[x];
             ix = glyphStorage.getCharIndex(x, success);
             glyphStorage.setCharIndex(x + 1, ix, success);
@@ -286,7 +286,7 @@ void IndicRearrangementProcessor::doRearrangementAction(LEGlyphStorage &glyphSto
         id = glyphStorage.getCharIndex(lastGlyph, success);
         x = lastGlyph - 2;
 
-        while (x > firstGlyph) {
+        while (x > (int32_t)firstGlyph) {
             glyphStorage[x + 1] = glyphStorage[x];
             ix = glyphStorage.getCharIndex(x, success);
             glyphStorage.setCharIndex(x + 1, ix, success);
@@ -316,7 +316,7 @@ void IndicRearrangementProcessor::doRearrangementAction(LEGlyphStorage &glyphSto
         id = glyphStorage.getCharIndex(lastGlyph, success);
         x = firstGlyph + 2;
 
-        while (x < lastGlyph) {
+        while (x < (int32_t)lastGlyph) {
             glyphStorage[x - 2] = glyphStorage[x];
             ix = glyphStorage.getCharIndex(x, success);
             glyphStorage.setCharIndex(x - 2, ix, success);
@@ -346,7 +346,7 @@ void IndicRearrangementProcessor::doRearrangementAction(LEGlyphStorage &glyphSto
         id = glyphStorage.getCharIndex(lastGlyph, success);
         x = firstGlyph + 2;
 
-        while (x < lastGlyph) {
+        while (x < (int32_t)lastGlyph) {
             glyphStorage[x - 2] = glyphStorage[x];
             ix = glyphStorage.getCharIndex(x, success);
             glyphStorage.setCharIndex(x - 2, ix, success);
